@@ -1,6 +1,6 @@
 module github.com/kubesphere/kubekey
 
-go 1.17
+go 1.18
 
 replace (
 	github.com/docker/distribution => github.com/docker/distribution v2.8.1+incompatible
