@@ -18,24 +18,23 @@ package v1beta1
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 )
-
-// EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
-// NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
 
 // KKMachineTemplateSpec defines the desired state of KKMachineTemplate
 type KKMachineTemplateSpec struct {
-	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
-	// Important: Run "make" to regenerate code after modifying this file
-
-	// Foo is an example field of KKMachineTemplate. Edit kkmachinetemplate_types.go to remove/update
-	Foo string `json:"foo,omitempty"`
+	Template KKMachineTemplateResource `json:"template"`
 }
 
-// KKMachineTemplateStatus defines the observed state of KKMachineTemplate
-type KKMachineTemplateStatus struct {
-	// INSERT ADDITIONAL STATUS FIELD - define observed state of cluster
-	// Important: Run "make" to regenerate code after modifying this file
+// KKMachineTemplateResource describes the data needed to create am AWSMachine from a template.
+type KKMachineTemplateResource struct {
+	// Standard object's metadata.
+	// More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
+	// +optional
+	ObjectMeta clusterv1.ObjectMeta `json:"metadata,omitempty"`
+
+	// Spec is the specification of the desired behavior of the machine.
+	Spec KKMachineSpec `json:"spec"`
 }
 
 //+kubebuilder:object:root=true
@@ -46,8 +45,7 @@ type KKMachineTemplate struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
 
-	Spec   KKMachineTemplateSpec   `json:"spec,omitempty"`
-	Status KKMachineTemplateStatus `json:"status,omitempty"`
+	Spec KKMachineTemplateSpec `json:"spec,omitempty"`
 }
 
 //+kubebuilder:object:root=true
