@@ -168,7 +168,6 @@ func (in *KKClusterList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *KKClusterSpec) DeepCopyInto(out *KKClusterSpec) {
 	*out = *in
-	in.Selector.DeepCopyInto(&out.Selector)
 	in.Nodes.DeepCopyInto(&out.Nodes)
 	out.ControlPlaneEndpoint = in.ControlPlaneEndpoint
 	in.Registry.DeepCopyInto(&out.Registry)
@@ -555,8 +554,7 @@ func (in *KKMachineTemplate) DeepCopyInto(out *KKMachineTemplate) {
 	*out = *in
 	out.TypeMeta = in.TypeMeta
 	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
-	out.Spec = in.Spec
-	out.Status = in.Status
+	in.Spec.DeepCopyInto(&out.Spec)
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KKMachineTemplate.
@@ -610,8 +608,26 @@ func (in *KKMachineTemplateList) DeepCopyObject() runtime.Object {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KKMachineTemplateResource) DeepCopyInto(out *KKMachineTemplateResource) {
+	*out = *in
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KKMachineTemplateResource.
+func (in *KKMachineTemplateResource) DeepCopy() *KKMachineTemplateResource {
+	if in == nil {
+		return nil
+	}
+	out := new(KKMachineTemplateResource)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *KKMachineTemplateSpec) DeepCopyInto(out *KKMachineTemplateSpec) {
 	*out = *in
+	in.Template.DeepCopyInto(&out.Template)
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KKMachineTemplateSpec.
@@ -620,21 +636,6 @@ func (in *KKMachineTemplateSpec) DeepCopy() *KKMachineTemplateSpec {
 		return nil
 	}
 	out := new(KKMachineTemplateSpec)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *KKMachineTemplateStatus) DeepCopyInto(out *KKMachineTemplateStatus) {
-	*out = *in
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KKMachineTemplateStatus.
-func (in *KKMachineTemplateStatus) DeepCopy() *KKMachineTemplateStatus {
-	if in == nil {
-		return nil
-	}
-	out := new(KKMachineTemplateStatus)
 	in.DeepCopyInto(out)
 	return out
 }
